Remove background color from inline mod menu.
The background color was previously forced to white, which looked fine
on normal posts (which have a white background), but was not correct
on deleted posts (which have a reddish-gray background). I presume
that it would also not be correct when the dark theme is used (although
I haven't tested it). I see no point in explicitly specifying a background
color here; it should just inherit its parent's background color.