Tweak appearance of "Load flagger stats" button.
Added the Stacks button styles to the "Load flagger stats" button
in order to make it align and blend in better.

I chose a filled standard button style, not a primary button, to make
the button visible, without being annoying. This is not a primary
button or one that performs an important action.