Resolve issue with floating user card in top bar
Following latest SE change to top bar the padding here is applied to the user card that appears in the top bar as well as the other user cards that appear on the page, pushing the users reputation, badges etc. of the top of the screen. Change disables padding-bottom css for the user card in the top bar.