Add "demands to show effort/'not a code-writing service'", "promotional content; answers not self-contained", and "generic year-long ban" templates
The "demands to show effort/'not a code-writing service'" template was
written by me.

"promotional content; answers not self-contained" was written by Makyen,
and is used with permission.

"generic year-long ban" is the same message used by
AdditionalInlinePostModMenu that is designed to fit within the notification
seen by a recreated account.