MMH: only add signature if explicit true addSignature in template
SE has changed to separately adding the signature to the resulting text,
rather than have it embedded in the text for each option. Thus, if we add
it, then there's two copies.

We should probably look at being able to remove it for `addSignature: false`,
but that's not implemented at this time.